+++ implementations/rust/ockam/ockam_command/src/tcp/outlet/create.rs
@@ -2,10 +2,9 @@ use crate::node::{default_node_name, node_name_parser};
 use crate::policy::{add_default_project_policy, has_policy};
 use crate::tcp::util::alias_parser;
 use crate::util::output::Output;
+use crate::util::parsers::socket_addr_parser;
 use crate::util::{extract_address_value, node_rpc, Rpc};
 use crate::CommandGlobalOpts;
-use crate::Result;
-use anyhow::anyhow;
 use clap::Args;
 use ockam::Context;
 use ockam_abac::Resource;
@@ -13,7 +12,7 @@ use ockam_api::nodes::models::portal::{CreateOutlet, OutletStatus};
 use ockam_core::api::{Request, RequestBuilder};
 
 use ockam_api::cli_state::{StateDirTrait, StateItemTrait};
-use std::net::{IpAddr, Ipv4Addr, SocketAddr};
+use std::net::SocketAddr;
 
 /// Create TCP Outlets
 #[derive(Clone, Debug, Args)]
@@ -39,40 +38,6 @@ impl CreateCommand {
     pub fn run(self, options: CommandGlobalOpts) {
         node_rpc(run_impl, (options, self))
     }
-}
-
-fn socket_addr_parser(input: &str) -> Result<SocketAddr> {
-    let to_address_info: Vec<&str> = input.split(':').collect();
-    if to_address_info.len() > 2 {
-        return Err(anyhow!("Failed to parse to address").into());
-    }
-
-    let port: u16 = if to_address_info.len() == 2 {
-        to_address_info[1]
-            .parse()
-            .map_err(|_| anyhow!("Invalid port number"))?
-    } else {
-        to_address_info[0]
-            .parse()
-            .map_err(|_| anyhow!("Invalid port number"))?
-    };
-
-    let server_ip: Ipv4Addr = if to_address_info.len() < 2 {
-        [127, 0, 0, 1].into()
-    } else {
-        let address_octets: [u8; 4] = {
-            let mut octets = [0; 4];
-            for (i, octet_str) in to_address_info[0].split('.').enumerate() {
-                octets[i] = octet_str
-                    .parse()
-                    .map_err(|_| anyhow!("Invalid IP address"))?;
-            }
-            octets
-        };
-        Ipv4Addr::from(address_octets)
-    };
-
-    Ok(SocketAddr::new(IpAddr::V4(server_ip), port))
 }
 
 fn default_from_addr() -> String {
